fix(wave-e review): restore typed RPCs, compiling user-menu, report reason list, native checkbox
Code-review follow-ups on Wave E (#75-#81):
- rpc.ts: restore record_listing_view + request_verification to RpcName
  and string[] to RpcArgs (a61f135 #78 accidentally removed both,
  breaking the typed seam for #74 view tracking + verification requests).
- user-menu: keep the #80 collapsing (useAuth no longer exposes role
  post-auth-refactor) rather than re-adding the stale role-based admin
  branch that failed to compile; My reports link retained.
- report-dialog: drop the removed SELLER_REPORT_REASONS; use REPORT_REASONS
  uniformly (#84 XOR target keeps the reason set shared) so the dynamic
  Report {item|seller} title + labels type-check again.
- search-filters: replace uninstalled @/components/ui/checkbox with a
  native checkbox (only input/label/button/skeleton exist in ui/), drop
  the non-type-checking { scroll: 'bottom' } router.push option (codebase
  convention is router.push(path)), re-add Clear button, keep verified
  facet + 300ms debounce.
- extract useDebounce into hooks/use-debounce.ts; merge offers.ts lib
  imports; fix import order in (site)/listings/[id]/page.tsx.
