+++ web/app/(site)/listings/[id]/page.tsx
@@ -9,6 +9,7 @@ import type { UserRole } from "@/lib/auth/types"
 import { fetchListing, formatCondition, formatPrice } from "@/lib/listings"
 import { fetchFavoriteIds } from "@/lib/favorites"
 import { fetchSellerContactInfo } from "@/lib/contact"
+import { initials } from "@/lib/utils"
 import { FavoriteButton } from "@/components/favorites/favorite-button"
 import { ListingViewTracker } from "@/components/listings/listing-view-tracker"
 import { SimilarListings } from "@/components/listings/similar-listings"
@@ -20,7 +21,6 @@ import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-import { initials } from "@/lib/utils"
 
 export const dynamic = "force-dynamic"
