+++ web/lib/offers.ts
@@ -3,8 +3,7 @@ import "server-only"
 import { createClient } from "@/lib/supabase/server"
 import type { Supabase } from "@/lib/supabase/types"
 import { callOutcomeRpc } from "@/lib/supabase/rpc"
-import { isValidUuid, PAGE_SIZE } from "@/lib/listings/constants"
-import type { BrowseListing } from "@/lib/listings/constants"
+import { isValidUuid, PAGE_SIZE, type BrowseListing } from "@/lib/listings/constants"
 import { mapNestedBrowseListing } from "@/lib/listings/browse-mapper"
 import type { NestedBrowseRow } from "@/lib/listings/browse-mapper"
 import {
