+++ web/lib/supabase/rpc.ts
@@ -10,13 +10,15 @@ export type RpcName =
   | "submit_report"
   | "resolve_report"
   | "record_verification"
+  | "request_verification"
+  | "record_listing_view"
   | "submit_offer"
   | "accept_offer"
   | "decline_offer"
   | "counter_offer"
   | "record_contact_attempt"
 
-export type RpcArgs = Record<string, string | number | boolean | null>
+export type RpcArgs = Record<string, string | number | boolean | null | string[]>
 
 export interface RpcResult<T> {
   data: T | null
